+++ Tema3/src/main/java/com/example/demo/dto/ProductTraderDto.java
@@ -11,14 +11,14 @@ public class ProductTraderDto {
     private String trader;
     private String productName;
     private Double price;
-    private Double quantity;
-    private Double maxQuantity;
-    private Double minQuantity;
+    private Long quantity;
+    private Long maxQuantity;
+    private Long minQuantity;
 
     public ProductTraderDto() {
     }
 
-    public ProductTraderDto(Long id, String producer, String trader, String product, Double price, Double quantity, Double maxQuantity, Double minQuantity) {
+    public ProductTraderDto(Long id, String producer, String trader, String product, Double price, Long quantity, Long maxQuantity, Long minQuantity) {
         this.id = id;
         this.producer = producer;
         this.trader = trader;
@@ -69,27 +69,27 @@ public class ProductTraderDto {
         this.price = price;
     }
 
-    public Double getQuantity() {
+    public Long getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Double quantity) {
+    public void setQuantity(Long quantity) {
         this.quantity = quantity;
     }
 
-    public Double getMaxQuantity() {
+    public Long getMaxQuantity() {
         return maxQuantity;
     }
 
-    public void setMaxQuantity(Double maxQuantity) {
+    public void setMaxQuantity(Long maxQuantity) {
         this.maxQuantity = maxQuantity;
     }
 
-    public Double getMinQuantity() {
+    public Long getMinQuantity() {
         return minQuantity;
     }
 
-    public void setMinQuantity(Double minQuantity) {
+    public void setMinQuantity(Long minQuantity) {
         this.minQuantity = minQuantity;
     }
 }
